LayoutTests/platform/ios-simulator/fast/dom/scroll-element-to-rect-expected.txt

Scrolled element to rect
FAIL rect.left should be 82. Was 610.
FAIL rect.top should be 139. Was 710.
PASS successfullyParsed is true

TEST COMPLETE
